+++ src/test/suite/extension.test.ts
@@ -6,7 +6,7 @@
 // The module 'assert' provides assertion methods from node
 import * as assert from 'assert';
 // import * as vscode from 'vscode'
-import { partitionString } from '../util';
+import { partitionString } from '../../util';
 
 
 // Defines a Mocha unit test
@@ -31,12 +31,13 @@ import { partitionString } from '../util';
 //     })
 //   }
 
+//to run tests start tsc -w and then go to the debug tab and select "EXtension Tests" and run
 
 // Defines a Mocha test suite to group tests of similar kind together
 suite("partitionString working properly", function () {
 
     test('partition size not fitting', async function () {
-        
+
         const text = '0123456789'
 
         const parts = partitionString(text, 3)
@@ -47,10 +48,10 @@ suite("partitionString working properly", function () {
     })
 
     test('partition size larger than text', async function () {
-        
+
         const text = '0123456789'
 
-        const parts = partitionString(text, text.length+10)
+        const parts = partitionString(text, text.length + 10)
 
         const margedText = parts.map(p => p.text).join('')
 
@@ -58,7 +59,7 @@ suite("partitionString working properly", function () {
     })
 
     test('partition size perfect fit', async function () {
-        
+
         const text = '0123456789'
 
         const parts = partitionString(text, text.length)
@@ -68,7 +69,52 @@ suite("partitionString working properly", function () {
         assert.equal(margedText, text)
     })
 
-});
 
+})
+
+
+suite('some frontend func tests', function () {
+
+    test('excel like column names (letters) func is correct (like handsontable)', async function () {
+
+        const COLUMN_LABEL_BASE = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
+        const COLUMN_LABEL_BASE_LENGTH = COLUMN_LABEL_BASE.length
+
+        function spreadsheetColumnLetterLabel(index: number) {
+            let num = index
+            let columnLabel = ''
+            //see https://stackoverflow.com/questions/34813980/getting-an-array-of-column-names-at-sheetjs
+            while (num >= 0) {
+                columnLabel = COLUMN_LABEL_BASE[num % 26] + columnLabel
+                num = Math.floor(num / 26) - 1
+            }
+            return columnLabel
+        }
+
+        function spreadsheetColumnLabel(index: number): string {
+            let dividend = index + 1
+            let columnLabel = ''
+            let modulo
+
+            while (dividend > 0) {
+                modulo = (dividend - 1) % COLUMN_LABEL_BASE_LENGTH;
+                columnLabel = String.fromCharCode(65 + modulo) + columnLabel;
+                dividend = parseInt((dividend - modulo) / COLUMN_LABEL_BASE_LENGTH as any, 10);
+            }
+
+            return columnLabel;
+        }
+
+
+        for (let i = 0; i < 1_000_000; i++) {
+            let correct = spreadsheetColumnLabel(i)
+            let test = spreadsheetColumnLetterLabel(i)
+
+            assert.equal(test, correct)
+        }
+
+    })
+
+})